add accept encoding header
Doesn't seem to really make much a performance difference. The main cause of waiting is still a handful of very slow responses from server (possibly due to rate limiting). So I need to shorten timeout on requests and retry and/or implement a queue system so things don't block as much.